EN-1577: Adds filter bar with option to all visualizations; (#575)
Visualizations can now be configured to show a filter bar! Simply
set $.fn.svgColumnChart(vif, options): options = { displayFilterBar:
true } and you're good to go.

This is set by default for embedded visualizations. Check out the
embed examples page at the bottom.

This also emits an event that VIF changed for future use of the
VIF that is mutated in SvgVisualization.

Version bumps to 21.5.0 and upgrades React and adds new peer dependency.