EN-15795: Other category omits null values (#584)
Core treats null values as if they don't exist if you throw a where
statement into the mix. For example, if you ask for the total number of
dogs in a dataset about dogs, our backend returns the expected number:
the total number of rows in the dataset. If you ask for the total
number of dogs that are not corgis and you happen to have a couple of
null values for the column we're asking about, Core will return an
unexpected number of rows: the number of matching rows, omitting any
null values. From Core's perspective, null means something does not
exist and it cannot distinguish it as being purposefully set to null.
While this is confusing, this is the known, expected behavior.

Since we cannot reliably ask for the number of rows that do not match a
where statement when we want to count null values as being values, this change:

- Doesn't put a limit on the request when showOtherCategory is true
- Checks to see if the rows we're going to put into the other category
  contain null values
- Adds 'OR <column> IS NOT NULL' if we're going to make a request for
  the other category and the rows that would fall into the other category
  contain null values

Because pie chart doesn't use the same data request code as any of the
other visualizations. I copy-pastaed essentially the same changes into its code.

I also updated the grouped column chart legends to handle null values.
It seems within reason that we might get null values from a grouped query,
but we hadn't handled that case previously.