ENG: Removes sdoc; We're not using it.
sdoc's dependency rdoc is causing a fuss right now:

Downloading rdoc-4.2.0 revealed dependencies not in the API or the
lockfile (json (~> 1.4)).
Either installing with `--full-index` or running `bundle update rdoc`
should fix the problem.

We don't use sdoc, so removing it is the quickest way to unblock us. As
for investigation, there are two theories according to bundler in debug mode:
a corrupt API or corrupt Gemfile.lock. Our Gemfile.lock hasn't change
significantly enough to warrant an error like this, so I'm going to lean
on some Artifactory repository issues. If the problem comes up again
elsewhere, that would be a good place to start.