Revert "EN-16532: consume phiddy metadata from common metadata methods (#4767)" (#4818)
This reverts commit 429c67f1b049b9334c7900e99af675394fa7bbbe.
The purpose of the commit was primarily to add logging in the ruby code to help us diagnose and fix the problem.
Fixing the issue with this commit was not a guarantee and we missed the bug in staging and rc.
(cherry picked from commit 68e4ad0ad58bada52d148e12c1266479bb02b058)