EN-16713: Use the transformed filename from core for upload-from-URL (#4888)
EN-16713: Use the transformed filename from core for upload-from-URL

Core will use the content-type of a downloaded file to determine what
to do with it during scanUrl, if it doesn't have an extension, and
return that information to frontend.

..for some reason we seem to have THREE versions of this code, and
it's the one named "old" that's actually used?  Anyway all three have,
I think, been updated to use this field.