EN-17490 - Socrata Viz Table incorrectly allows sorting by blob type
The function that was being used to determine whether or not to allow sorting by a column.

This was problematic because blob columns are also not sortable, but we were allowing the user to attempt to sort by them, which made the resulting query fail in such a way as to put the table into an unrecoverable broken state.