EN-17394: Remove absolute positioning (#5147)
According to the NASA 508 compliance team, using absolute positioning is a no-no, so removing it here. The effect is that the "* required" text appears below the form, above the modal buttons, instead of to the left of the buttons.
(cherry picked from commit 0ab4627aeb601c790316eda5da7ef37011b1facd)