EN-17639: Use consistent TZ for date formatting
Switching to the momemt-timezone module to allow us to be able to use the moment.tz() function, which is added by this module. The reason we want to this is to try and normalize how we convert UTC dates to the user-local timezone and format them, with the aim of presenting all dates in the UI in a consistent manner.

In the ideal world we would not be duplicating code like this. Enabling all-screens.js to import dateLocalize from locale.js would involve changing webpack to run all of the old legacy JS code through the import, babel, and other pipelines, which is simply too large a change to take on at present.