EN-17277: Introduce horrible 'deoptimizeRender' workaround for legacy viz in stories
There's existing code in some legacy visualizations that determines whether or not
to render pieces of the chart based on whether or not those pieces are considered
visible within the viewport. It is highly beneficial when DSG is applied, because
otherwise the performance in some browsers (e.g. Firefox) can be badly degraded.
However, it has a poor interaction with Storyteller's resizable rich-media blocks,
causing axis labels and data value bars/columns to be dropped from the rendered
output when the block is sufficiently tall, even though the whole chart area is
actually visible.

We generally don't see the types of large, complex DSG visualizations associated
with degraded performance in a Storyteller context, so as a workaround this change
allows us to short-circuit the "within-viewport" detection logic, thereby treating
everything as renderable. This approach is pretty terrible, but not more terrible
than spending lots of time making a riskier change to adjust the affected logic
in a generation of visualization components that we're trying to retire.

See 34c8b35fbf48b5fd48abf37ef9fea83fe0181912 for the original related work.