EN-17879: Omit working copies with datasets filter (#5210)
EN-17879: Omit working copies with datasets filter

Now that we have a working copies filter, when we filter on datasets, we need to omit the working copies from that set.

When we filter on "working copies" we are really filtering on "datasets" with the added constraint of "published=true". This change provides access to the filter state so we can change the count label accordingly.