EN-9125: Move Jenkins build scripts from jobs to source control
The build scripts for frontend currently live in the job configuration
and are run inside containers using a cloud plugin based on Docker. As
we try and upgrade Docker on the Jenkins machine, the plugin providing
the cloud is no longer compatible with the newest Docker and doesn't
seem to be developed anymore. The build scripts in Jenkins have now
mostly been replaced with calls to `docker run` using these scripts as
the commands to execute. This allows us to move forward with the Docker
upgrade and leave the old Docker plugin behind.