EN-16762: Fix roles API URL (#5385)
When running in a deployed environment, the API URL was being set to
localhost, causing API requests to try to be performed as CORS requests,
and failing.  This change ensures that the API base URL is the current
domain.
QA: The /admin/roles page will not display an error, and instead be usable.