EN-16569: Improve chunking implementation and tests.
1. Reduce the limit to 1400 from 2000. The maximum request length, empirically,
seems to be ~3000 bytes. We may have up to 2 long IN clauses in the
GroupedCategoricalDataManager world, so giving each half the available length,
with another ~200 bytes for other query overhead, should be reasonable.

2. Chunking needs to use bytes, not true string length. This was empirically
determined by hitting the core request server, finding its length limit, and
replacing one 1-byte character with a 2-byte character and finding that it
chokes. The chunking implementation now supports both possibilities.