EN-18411: Address build breakage caused by lint changes.
Root cause: Node resolves modules _relative to the currently-executing
file_. This doesn't work well if the file in question is under common/
but frontend/ is requiring it. That means that everything that the file
in question tries to require() is resolved relative to common/, which
does not have node_modules. Ultimately it will fall back to your
system-wide modules, which won't work unless you've installed everything
globally.

Separately, eslint really wants you to name your shared configs
eslint-config-*.

Fix is to add a NODE_PATH to force node to always at least consider
frontend's node_modules when running within frontend.