EN-19802-adjacent: Stop special-casing href datasets' urls' keys
This is not a feature that's been used in years, and never on domains
of any current customer (it was data.gov-specific code).  Denormalizing
the internet like this was doing is probably a bad idea; if we ever need
this sort of thing again it'd probably be better to make HEAD requests
on-demand to display the sizes of things linked to.