EN-19811: default to parse source false for unparsable extensions (#5849)
EN-19811: dsmui: unparsable extensions create parse_source: false sources

This is a simple change that tells DSMUI to create a source with parse_source false from the start, if the file being uploaded doesn't have a supported extension. A more in-depth solution is in progress to account for handling extensions that are ambiguously parseable (.json, .zip). Since those extensions are both geospatial, this is also nice as it prevents the geo-spatial path if the feature flag is not on.

Bonuses:
* removes unused dispatch from TablePreviewContainer.js
* fix null content type bug