EN-20516: Fix styles for viz-can "more info" button
It appears that the bug was introduced with 72233bd892f8, but it's
unclear to me why. Basically what happened was that the $light-grey-8
variable was not getting the tint() function applied correctly and
so it had been defaulting to the $grey value. Some change caused tint()
to start working again, making the color of $light-grey-8 actually
a very light color.