EN-20923: Fix rendering filtered derived views
A recent change to core (https://github.com/socrata/core/commit/d089dd1e28b0f1bbeb9e5acc30c28d38cad1d44e)
made it so that requesting federated views no longer responds
with 404 without federation headers. In the code that requests
the parent view, we were not sending headers. We are also not
handling the redirect that now occurs for this request. The fix
here was to start sending federation headers with the request.