EN-17028: First Pass Colocation UI (#6195)
EN-17028: Created basic UI to allow users to colocate 2 datasets. Does not have a route to get to, just need to enter the path directly. This is hidden behind a feature flag for the moment, users should not be able to hit it. Colocation is still in development so this experience will likely be changing pretty drastically.