EN-20524: Make vizcan edit bar not suck on small devices
We do not want to show all the creation
buttons on the small screen sizes for
now since the AX is not designed for
those small sizes. Instead, we let users
filter and save datasets in viz can.
Later, we may choose to do more with
mobile responsiveness for this surface.