EN-21341 Allow skipping of localeapp and tests
This is only meant for use in the frontend-canary build. Because
Hai’s team has separate testers the long frontend build process has
been keeping them from verifying their work in a timely manner.
(Frontend builds are up to ~2 hours). So allow them to skip tests and
the locale app push/pull (which often takes 30-60 minutes on it’s own).