EN-21473: Fix remote code execution vulnerability in RemotePartialsController.
It is possible to inject arbitrary-ish ruby code to be executed via
parameters to RemotePartialsController.

eg. https://localhost/templates/page_render_type?File.write(%22/tmp/marcswuzhere.txt%22,%20%22%22)||bar

This is due to a bug in rails where it puts the key into a string
of ruby code that it then evals:
https://github.com/rails/rails/blob/7847a19f476fb9bee287681586d872ea43785e53/actionview/lib/action_view/template.rb#L316

While the real fix is in rails, Sumo searches have shown we don't
actually pass parameters to these in practice so the simplest
solution is to just not allow passing them.

Rails 5.1.0 and later should avoid this issue due to a change
made for related but non-security reasons in commit
https://github.com/rails/rails/commit/f9960f2d74b510e0d994d49377cec36301f7e7f0