EN-21775: Logic for nbe_bucket_size flag was broken.
`nbe_bucket_size` is an old feature flag that does fancy logic where
`true` and `false` are meaningful and preferred, even though it
supports integers. It's a bit of cleverness that I now know better than
to try in our codebase, and wish I hadn't done.

The breaking commit is here:
https://github.com/socrata/platform-ui/commit/e97bea753d2#diff-aeac3ac4f1ea8f9d386879bdd44ab97bL131