EN-21839: Fix wacky spinning button text in storyteller
A change to the styleguide styles for buttons has caused our buttons in
storyteller to do a funky spin thing. Updated the button in the action bar
to the React <Button> component. Stuck a bandaid fix on the button
in the settings panel because I expect that code to go away when we
add the Action bar to storyteller. It's also not very conducive to adding
a react component the way it's currently written.