EN-22002: Viz can visualizations link to base dataset (#6550)
A previous change made it so that "View source data" links use the
`origin` property of the vif. Saved viz-can visualizations have this
property so it ends up making the link point back to the current page.
This makes it so that it behaves like it did before, where they link to
the base dataset when rendered within a viz-canvas.