EN-21490: Fix classic viz preview in storyteller
With the recent changes to use the common asset selector in storyteller,
we started to use the common `MetadataProvider` code to get the view for
the classic visualization in order to preview it and add it to the
story. Well, the `MetadataProvider.getDatasetMetadata()` method appends
`?read_from_nbe=true&version=2.1` to the metadata request, which returns
unexpected results in the view. The change here is to not use those nbe
flags when getting the view for rendering the visualizations.